refactor: コードの可読性を高めるためのリファクタ
